requirements.txt
numpy== 1.22.3 
scikit-learn==1.2.2
matplotlib==3.7.5
gensim==4.3.3 
scipy==1.7.3
pandas==1.5.3
seaborn==0.13.2
arabic-reshaper==3.0.0 
python-bidi==0.6.0 
mlxtend==0.23.1